Use MoK variable fallback
Shim is supposed to create an EFI table containing the MoK configuration which
the kernel accesses via the efi_mokvar_* functions. In some cases (possibly due
to Xen), this table is not available. The configuration is also made available
via EFI variables and the kernel has fallback paths that use the variables if
the table is not available. Fall back to using the EFI variable in this case
too.

With this change, MoK keys correctly get added to the machine keying which is
trusted and can be used to verify modules (rather than the untrusted platform
keyring).

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>