Revert to use num_online_cpus for default rss queues
In kernel version v4.19.19, the default number of RSS queues is
determined using num_online_cpus(). However, in kernel v6.6.x, the logic
is updated to use CPU topology information to determine the number of
physical cores.

For Xen dom0 kernels, this change introduces a problem: the dom0 kernel
cannot retrieve valid CPU topology data from Xen. As a result, the
default number of RSS queues is always set to 1, which degrades
networking performance of guests using xen-netback.

This patch introduces a fallback mechanism: if CPU topology information
is unavailable, it reverts to using num_online_cpus(). To avoid excessive
queue allocation, the number of RSS queues is capped at 8.

Signed-off-by: Ming Lu <ming.lu@cloud.com>